Add files via upload
This is a fix that will prevent the CI/CD from failing the deployment. However, the whole Quizzy folder should probably be rid of as we have already restructured the code.
Option 1:
Upload this and just keep going
Option 2:
Just delete the original Quizzy folder as restructing has already happened.